requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
# ecdsa
# paramiko
# pycrypto